Removed compilation warning.
git-svn-id: http://seriss.com/public/fltk/fltk/branches/branch-1.3-Fl_Printer@7192 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/Fl_Quartz_Printer.mm
@@ -259,7 +259,7 @@ void Fl_Quartz_Printer::end_job (void)
   
   status = PMSessionError(printSession);
   if (status != noErr) {
-    fl_alert ("PM Session error %d", status);
+    fl_alert ("PM Session error %d", (int)status);
   }
   PMSessionEndDocumentNoDialog(printSession);
   fl_gc = 0;
